requirements.txt
<<<<<<< HEAD
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0

=======
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0

>>>>>>> 966fed7 (initial push)
